+++ packages/engine-multi/src/index.ts
@@ -3,3 +3,4 @@ import createEngine from './api';
 export default createEngine;
 
 export * from './types';
+export * from './events';