+++ packages/lightning-mock/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @openfn/lightning-mock
 
+## 1.0.4
+
+### Patch Changes
+
+- 06f2f75: Acknowledge attempt start
+- Updated dependencies
+  - @openfn/engine-multi@0.1.3
+
 ## 1.0.3
 
 ### Patch Changes